PersonMapping sınıfı isim düzenlemesi. Hatalı test methodunun çıkartılması.
